+++ src/Products/PloneMeeting/MeetingItem.py
@@ -88,7 +88,9 @@ from Products.PloneMeeting.config import READER_USECASES
 from Products.PloneMeeting.config import REINDEX_NEEDED_MARKER
 from Products.PloneMeeting.config import SENT_TO_OTHER_MC_ANNOTATION_BASE_KEY
 from Products.PloneMeeting.config import WriteBudgetInfos
+from Products.PloneMeeting.config import WriteDecision
 from Products.PloneMeeting.config import WriteInternalNotes
+from Products.PloneMeeting.config import WriteItemMeetingManagerFields
 from Products.PloneMeeting.config import WriteMarginalNotes
 from Products.PloneMeeting.content.meeting import Meeting
 from Products.PloneMeeting.events import item_added_or_initialized
@@ -1296,7 +1298,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     TextField(
         name='decision',
@@ -1311,7 +1313,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=False,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     TextField(
         name='decisionSuite',
@@ -1327,7 +1329,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     TextField(
         name='decisionEnd',
@@ -1343,7 +1345,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     BooleanField(
         name='oralQuestion',
@@ -1399,7 +1401,7 @@ schema = Schema((
         default_content_type="text/html",
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='notes',
@@ -1415,7 +1417,7 @@ schema = Schema((
         default_content_type="text/html",
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='meetingManagersNotes',
@@ -1431,7 +1433,7 @@ schema = Schema((
         default_content_type="text/html",
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='meetingManagersNotesSuite',
@@ -1447,7 +1449,7 @@ schema = Schema((
         default_content_type="text/html",
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='meetingManagersNotesEnd',
@@ -1463,7 +1465,7 @@ schema = Schema((
         default_content_type="text/html",
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='internalNotes',
@@ -1514,7 +1516,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     LinesField(
         name='templateUsingGroups',
@@ -1665,7 +1667,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='committeeObservations',
@@ -1681,7 +1683,7 @@ schema = Schema((
         default_output_type="text/x-html-safe",
         searchable=True,
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     TextField(
         name='votesObservations',
@@ -1697,7 +1699,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
     ),
     ReferenceField(
         name='manuallyLinkedItems',
@@ -1817,7 +1819,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     TextField(
         name='otherMeetingConfigsClonableToFieldDecision',
@@ -1833,7 +1835,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     TextField(
         name='otherMeetingConfigsClonableToFieldDecisionSuite',
@@ -1849,7 +1851,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     TextField(
         name='otherMeetingConfigsClonableToFieldDecisionEnd',
@@ -1865,7 +1867,7 @@ schema = Schema((
         allowable_content_types=('text/html',),
         default_output_type="text/x-html-safe",
         optional=True,
-        write_permission="PloneMeeting: Write decision",
+        write_permission=WriteDecision,
     ),
     BooleanField(
         name='isAcceptableOutOfMeeting',
@@ -1954,7 +1956,7 @@ schema = Schema((
             i18n_domain='PloneMeeting',
         ),
         optional=True,
-        write_permission="PloneMeeting: Write item MeetingManager reserved fields",
+        write_permission=WriteItemMeetingManagerFields,
         default_output_type="text/x-html-safe",
         default_content_type="text/plain",
     ),